feat: change operation graph API
