+++ pnpxai/explainers/rap/rap.py
@@ -1,8 +1,7 @@
-from typing import List, Dict
-from collections import defaultdict
+from typing import Dict
 
 import torch
-from torch import nn, Tensor, fx
+from torch import nn, Tensor
 from pnpxai.explainers.rap import rules
 from pnpxai.explainers.utils.operation_graph import OperationGraph, OperationNode
 
@@ -17,7 +16,6 @@ SUPPORTED_MODULES: Dict[type[nn.Module], type[rules.RelProp]] = {
     nn.Linear: rules.Linear,
     nn.Conv2d: rules.Conv2d,
 }
-
 SUPPORTED_FUNCTIONS: Dict[callable, type[rules.RelProp]] = {
     torch.add: rules.Add,
     torch.flatten: rules.Flatten,
@@ -37,7 +35,7 @@ class RelativeAttributePropagation():
 
     def _assign_rules_and_hooks(self, node: OperationNode):
         if node.is_module:
-            layer = node.method
+            layer = node.operator
             if type(layer) in SUPPORTED_MODULES and not (hasattr(layer, 'rule')):
                 rule = SUPPORTED_MODULES[type(layer)]
                 layer.rule: rules.RelProp = rule(layer)
@@ -47,55 +45,18 @@ class RelativeAttributePropagation():
             if not next_node.is_output:
                 self._assign_rules_and_hooks(next_node)
 
-    # def relprop(self, r: Tensor) -> Tensor:
-    #     def _relprop(node: OperationNode):
-    #         if node.is_output:
-    #             return r.clone()
-
-    #         cur_relprops = []
-    #         for next_node in node.next_nodes:
-    #             method = node.method
-    #             next_relprop = None
-    #             args_list = [
-    #                 prev_node.method.rule.Y for prev_node in node.prev_nodes if prev_node.is_module]
-    #             args = args_list[0] if len(args_list) == 1 else args_list
-    #             rule = None
-    #             if node.is_placeholder:
-    #                 rule = rules.RelProp()
-    #             elif node.is_module and type(method) in SUPPORTED_MODULES:
-    #                 rule = method.rule
-    #             elif node.is_function:
-    #                 if type(method) in SUPPORTED_FUNCTIONS:
-    #                     rule = SUPPORTED_FUNCTIONS[type(method)](method)
-    #                 else:
-    #                     built_in_name = str(method)[1:-1].split(' ')
-    #                     if len(built_in_name) >= 3 and built_in_name[2] in SUPPORTED_BUILTINS:
-    #                         rule = SUPPORTED_BUILTINS[built_in_name[2]]()
-
-    #             if rule is None:
-    #                 raise NotImplementedError(f'Unsupported node: {node}')
-    #             print("BEFORE: ", node, rule)
-    #             next_relprop = _relprop(next_node)
-    #             print("AFTER: ", node, rule)
-    #             cur_relprop = rule.relprop(next_relprop, args)
-    #             cur_relprops.append(cur_relprop)
-
-    #         return sum(cur_relprops) if len(cur_relprops) > 1 else cur_relprops[0]
-
-    #     return _relprop(self.graph.root)
-
-    def get_node_rule(self, node) -> rules.RelProp:
-        method = node.method
+    def get_node_rule(self, node: OperationNode) -> rules.RelProp:
+        operator = node.operator
         rule = None
         if node.is_placeholder:
             rule = rules.RelProp()
-        elif node.is_module and type(method) in SUPPORTED_MODULES:
-            rule = method.rule
+        elif node.is_module and type(operator) in SUPPORTED_MODULES:
+            rule = operator.rule
         elif node.is_function:
-            if type(method) in SUPPORTED_FUNCTIONS:
-                rule = SUPPORTED_FUNCTIONS[type(method)](method)
+            if type(operator) in SUPPORTED_FUNCTIONS:
+                rule = SUPPORTED_FUNCTIONS[type(operator)](operator)
             else:
-                built_in_name = str(method)[1:-1].split(' ')
+                built_in_name = str(operator)[1:-1].split(' ')
                 if len(built_in_name) >= 3 and built_in_name[2] in SUPPORTED_BUILTINS:
                     rule = SUPPORTED_BUILTINS[built_in_name[2]]()
 
@@ -116,7 +77,7 @@ class RelativeAttributePropagation():
                 stack[preserved_node] = preserved_rs
 
             args_list = [
-                prev_node.method.rule.Y
+                prev_node.operator.rule.Y
                 for prev_node in node.prev_nodes
                 if prev_node.is_module
             ]
