+++ pnpxai/explainers/utils/operation_graph.py
@@ -2,9 +2,9 @@ from torch import fx, nn
 from typing import Optional, List, Union, Callable, Dict
 
 
-def get_model_method(model: nn.Module, method_name: str):
-    method_path = method_name.split('.')
-    for block in method_path:
+def get_model_operator(model: nn.Module, operator_name: str):
+    operator_path = operator_name.split('.')
+    for block in operator_path:
         model = model._modules[block]
     return model
 
@@ -13,12 +13,12 @@ class OperationNode:
     def __init__(
         self,
         node: fx.Node,
-        method: Optional[Union[nn.Module, Callable]] = None,
+        operator: Optional[Union[nn.Module, Callable]] = None,
         next_nodes: Optional[List["OperationNode"]] = None,
         prev_nodes: Optional[List["OperationNode"]] = None,
     ):
         self._node = node
-        self.method = method
+        self.operator = operator
         self._next_nodes = next_nodes or []
         self._prev_nodes = prev_nodes or []
 
@@ -28,10 +28,10 @@ class OperationNode:
     def add_prev_node(self, node: "OperationNode"):
         self._prev_nodes.append(node)
 
-    def set_method_from_model(self, model: nn.Module):
+    def set_operator_from_model(self, model: nn.Module):
         if self.is_module or self.is_function:
             target = self._node.target
-            self.method = get_model_method(model, target) \
+            self.operator = get_model_operator(model, target) \
                 if isinstance(target, str) else target
 
     @property
@@ -88,7 +88,7 @@ class OperationGraph:
 
         for node in graph.nodes:
             operation_node = OperationNode(node)
-            operation_node.set_method_from_model(self._model)
+            operation_node.set_operator_from_model(self._model)
             self._nodes_map[node.name] = operation_node
             prev_nodes = node.args
